+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Created MPButtonTypes chooseTHConfigAndRunTherion and runTherion to run Therion.
   * Created TH2FileEditcontroller.getTHFileWidgetBuildContext().
   * Removed unused methods from TH2FileEditElementEditController.
+  * Created addPoint() method in TH2FileEditElementEditController.
 
 ## 0.2.38 - 2026-06-01 - The [Run therion in Flathub] release